requirements.txt
pyfiglet==0.8.post1
jina[perf]==3.13.2.dev34
jina-hubble-sdk>=0.15.5
pyfiglet==0.8.post1
cowsay==4.0
yaspin==2.1.0
py-cpuinfo==8.0.0
kubernetes==23.3.0
prompt_toolkit==3.0.28
docarray[common]==0.19.1
requests==2.27.1
pydub==0.25.1
nltk==3.7
jcloud==0.1.6
python-dotenv
boto3==1.26.43
elasticsearch==8.4.1
pydantic==1.9.2
httpx==0.23.2